Most updated stuff
Been working on it all and it feels like some stuff is breaking more than it's working right now. Just giving you guys the most updated version of what's going on at my end. If you want to take a crack at it then go for it. The Griffin is no longer following the player and I can't get the state sprites to change like I want.